# openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beyond Auth",
    "description": "Per-project authentication and authorization service.",
    "license": {
      "name": "AGPL-3.0",
      "identifier": "AGPL-3.0"
    },
    "version": "1"
  },
  "paths": {
...
    "/v1/passkey-registrations": {
      "post": {
        "tags": [
          "passkeys"
        ],
        "summary": "Begin WebAuthn passkey registration. Returns a `options` object to pass directly to\n`navigator.credentials.create()` and a `state_token` to include in the subsequent\n`POST /v1/passkeys` finish call. Two-step: begin here, finish with `POST /v1/passkeys`.",
        "operationId": "begin_passkey_registration",
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeginResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/passkeys": {
      "get": {
        "tags": [
          "passkey"
        ],
        "summary": "List all registered passkey credentials for the authenticated user.",
        "operationId": "list_passkeys",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CredentialRecord"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "passkeys"
        ],
        "summary": "Complete passkey registration. Submit the `state_token` from `POST /v1/passkey-registrations`\nand the `PublicKeyCredential` response from `navigator.credentials.create()`.\nThe registered credential can then be used to authenticate via `POST /v1/passkey-authentications`.",
        "operationId": "create_passkey",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinishRegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredCredential"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/v1/passkeys/{id}": {
      "delete": {
        "tags": [
          "passkeys"
        ],
        "summary": "Delete a registered passkey credential. The credential can no longer be used to authenticate.",
        "operationId": "delete_passkey",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Credential ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Credential deleted"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "patch": {
        "tags": [
          "passkeys"
        ],
        "summary": "Update the nickname of a registered passkey credential.",
        "operationId": "update_passkey",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Credential ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCredentialRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Credential updated"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/v1/sessions": {
      "get": {
        "tags": [
          "sessions"
        ],
        "summary": "List all active sessions for the authenticated user. The current session is included.",
        "operationId": "list_sessions",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionsResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "sessions"
        ],
        "summary": "Authenticate and create a session. The `grant_type` field in the request body selects\nthe credential flow. On success, returns 201 with a session token. When the account has\nTOTP enrolled and the grant type supports MFA, returns 200 with a `step_up_token` —\nthe caller must then re-POST with `grant_type=totp_step_up` to complete authentication.",
        "operationId": "create_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "MFA step-up required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StepUpResponse"
                }
              }
            }
          },
          "201": {
            "description": "Session created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Password validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "sessions"
        ],
        "summary": "Revoke all sessions for the authenticated user. Use `except_current=true` to keep\nthe current session active (e.g. \"sign out everywhere else\").",
        "operationId": "delete_all_sessions",
        "parameters": [
          {
            "name": "except_current",
            "in": "query",
            "description": "When true, the current session is preserved; all other sessions are revoked.\nDefaults to false — all sessions including the current one are revoked.",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Sessions revoked"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "type": "object",
        "description": "Returned on successful login or signup. Contains the new session alongside the user,\nprimary email, and personal org context.",
        "required": [
          "user",
          "email",
          "org",
          "session"
        ],
        "properties": {
          "email": {
            "$ref": "#/components/schemas/EmailBody"
          },
          "org": {
            "$ref": "#/components/schemas/OrgBody"
          },
          "session": {
            "$ref": "#/components/schemas/SessionBody"
          },
          "user": {
            "$ref": "#/components/schemas/UserBody"
          }
        }
      },
...
      "BeginResponse": {
        "type": "object",
        "required": [
          "options",
          "state_token"
        ],
        "properties": {
          "options": {
            "type": "object",
            "description": "WebAuthn options object — pass directly to the browser's WebAuthn API."
          },
          "state_token": {
            "type": "string",
            "description": "Opaque state token; include in the corresponding finish request."
          }
        }
      },
...
      "CredentialRecord": {
        "type": "object",
        "required": [
          "id",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "last_used_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "nickname": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "EmailBody": {
        "type": "object",
        "description": "Primary email address for the user.",
        "required": [
          "id",
          "email"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "verified_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Null if the address has not been verified."
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "Wire-format error body returned on all non-2xx responses.",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable error code, e.g. `\"invalid_credentials\"`."
          },
          "hint": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional actionable guidance present on configuration-gate errors."
          },
          "message": {
            "type": "string",
            "description": "Human-readable description."
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "Top-level error envelope for all API error responses.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
      "FinishRegistrationRequest": {
        "type": "object",
        "required": [
          "state_token",
          "credential"
        ],
        "properties": {
          "credential": {
            "type": "object",
            "description": "WebAuthn `PublicKeyCredential` response from the browser."
          },
          "nickname": {
            "type": [
              "string",
              "null"
            ]
          },
          "state_token": {
            "type": "string"
          }
        }
      },
...
      "LoginRequest": {
        "oneOf": [
          {
            "type": "object",
            "description": "Email + password login. If TOTP is enrolled, returns 200 with a `step_up_token`\ninstead of a session — caller must complete MFA via `totp_step_up`.",
            "required": [
              "email",
              "password",
              "grant_type"
            ],
            "properties": {
              "email": {
                "type": "string"
              },
              "grant_type": {
                "type": "string",
                "enum": [
                  "password"
                ]
              },
              "password": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "description": "Exchanges a token issued by `POST /v1/magic-links`. Single-use; expires in 15 min.",
            "required": [
              "token",
              "grant_type"
            ],
            "properties": {
              "grant_type": {
                "type": "string",
                "enum": [
                  "magic_link"
                ]
              },
              "token": {
                "type": "string",
                "description": "The plaintext token from the magic-link response."
              }
            }
          },
          {
            "type": "object",
            "description": "Exchanges a password-reset token, sets the new password, and creates a session.\nAll existing sessions are invalidated.",
            "required": [
              "token",
              "new_password",
              "grant_type"
            ],
            "properties": {
              "grant_type": {
                "type": "string",
                "enum": [
                  "password_reset"
                ]
              },
              "new_password": {
                "type": "string"
              },
              "token": {
                "type": "string",
                "description": "The plaintext token from the password-reset response."
              }
            }
          },
          {
            "type": "object",
            "description": "Exchanges an email-change token, promotes the new address to primary, and creates\na session. All existing sessions are invalidated.",
            "required": [
              "token",
              "grant_type"
            ],
            "properties": {
              "grant_type": {
                "type": "string",
                "enum": [
                  "email_change"
                ]
              },
              "token": {
                "type": "string",
                "description": "The plaintext token from the email-change (`POST /v1/emails`) response."
              }
            }
          },
          {
            "type": "object",
            "description": "Completes a TOTP MFA step-up. Present after a 200 response with `step_up_required`.",
            "required": [
              "step_up_token",
              "code",
              "grant_type"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "Six-digit TOTP code from the authenticator app."
              },
              "grant_type": {
                "type": "string",
                "enum": [
                  "totp_step_up"
                ]
              },
              "step_up_token": {
                "type": "string",
                "description": "The `step_up_token` from the preceding 200 step-up response."
              }
            }
          },
          {
            "type": "object",
            "description": "Completes MFA using a single-use TOTP recovery code. Consumes the code permanently.",
            "required": [
              "step_up_token",
              "code",
              "grant_type"
            ],
            "properties": {
              "code": {
                "type": "string",
                "description": "One of the recovery codes issued at TOTP enrollment or last regeneration."
              },
              "grant_type": {
                "type": "string",
                "enum": [
                  "totp_recovery"
                ]
              },
              "step_up_token": {
                "type": "string",
                "description": "The `step_up_token` from the preceding 200 step-up response."
              }
            }
          },
          {
            "type": "object",
            "description": "Completes a WebAuthn passkey authentication flow.",
            "required": [
              "state_token",
              "credential",
              "grant_type"
            ],
            "properties": {
              "credential": {
                "type": "object",
                "description": "WebAuthn `PublicKeyCredential` response from the browser's `navigator.credentials.get()`."
              },
              "grant_type": {
                "type": "string",
                "enum": [
                  "passkey"
                ]
              },
              "state_token": {
                "type": "string",
                "description": "The `state_token` from the `POST /v1/passkey-authentications` response."
              }
            }
          }
        ],
        "description": "Credential payload for session creation. The `grant_type` field discriminates the variant."
      },
...
      "OrgBody": {
        "type": "object",
        "description": "The user's personal org.",
        "required": [
          "id",
          "name",
          "slug"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "image_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string",
            "description": "URL-safe identifier, unique across all orgs."
          }
        }
      },
...
      "RegisteredCredential": {
        "type": "object",
        "required": [
          "id",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nickname": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "SessionBody": {
        "type": "object",
        "description": "Session credential returned on login or signup.",
        "required": [
          "id",
          "token",
          "expires_at"
        ],
        "properties": {
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "type": "string",
            "description": "Opaque bearer token — store securely, transmit as `Authorization: Bearer <token>`."
          }
        }
      },
      "SessionListItem": {
        "type": "object",
        "required": [
          "id",
          "token_id",
          "created_at",
          "expires_at",
          "current"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "current": {
            "type": "boolean",
            "description": "Whether this is the caller's current session."
          },
          "expires_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "ip_address": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_used_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "token_id": {
            "type": "string",
            "format": "uuid"
          },
          "user_agent": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "SessionsResponse": {
        "type": "object",
        "description": "Paginated list of active sessions for the authenticated user.",
        "required": [
          "sessions"
        ],
        "properties": {
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SessionListItem"
            }
          }
        }
      },
...
      "StepUpKind": {
        "type": "string",
        "description": "MFA method required to complete authentication.",
        "enum": [
          "totp"
        ]
      },
      "StepUpResponse": {
        "type": "object",
        "description": "Returned when the user has TOTP enrolled — caller must complete the step-up flow.",
        "required": [
          "step_up_required",
          "step_up_token"
        ],
        "properties": {
          "step_up_required": {
            "$ref": "#/components/schemas/StepUpKind",
            "description": "The MFA method required to complete authentication."
          },
          "step_up_token": {
            "type": "string",
            "description": "Short-lived signed token to present when completing the step-up."
          }
        }
      },
...
      "UpdateCredentialRequest": {
        "type": "object",
        "description": "Request to rename a registered passkey credential.",
        "required": [
          "nickname"
        ],
        "properties": {
          "nickname": {
            "type": "string",
            "description": "Human-readable label shown in credential lists."
          }
        }
      },
...
      "UserBody": {
        "type": "object",
        "description": "Core user fields.",
        "required": [
          "id",
          "primary_org_id",
          "name",
          "metadata",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "image_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "metadata": {
            "type": "object",
            "description": "Arbitrary JSON metadata stored on the personal org."
          },
          "name": {
            "type": "string",
            "description": "Display name (mirrors the personal org name)."
          },
          "primary_org_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the user's personal org. Always present; created at signup."
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "session_<id>_<secret>"
      }
    }
  },
...
}